requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
Werkzeug==0.12.2
